# docs/swagger.yaml
definitions:
  domain.Dentist:
    properties:
      license:
        type: string
      id:
        type: integer
      lastname:
        type: string
      name:
        type: string
    type: object
  domain.Patient:
    properties:
      dateup:
        type: string
      dni:
        type: string
      address:
        type: string
      id:
        type: integer
      lastname:
        type: string
      name:
        type: string
    required:
    - dateup
    - dni
    type: object
  dto.appointmentInsert:
    properties:
      dateup:
        type: string
      dentistid:
        type: integer
      description:
        type: string
      hour:
        type: string
      patientid:
        type: integer
    required:
    - dateup
    - dentistid
    - hour
    - patientid
    type: object
  dto.appointmentPost:
    properties:
      dateup:
        type: string
      description:
        type: string
      dni:
        type: string
      license:
        type: string
      hour:
        type: string
    type: object
  web.response:
    properties:
      data: {}
    type: object
info:
  contact:
    name: API Support jimena_shk.
    url: https://github.com/jimena_shk
  description: API of a dental clinic, where you can list/ add/ delete/ update or search by ID Patients, Dentists and Appointments.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/jimena_shk
  title: Api Dental Clinic
  version: "1.0"
paths:
  /dentists:
    get:
      consumes:
      - application/json
      description: Get Dentist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
      summary: Get All Dentist
      tags:
      - Dentist
    post:
      consumes:
      - application/json
      description: Post a Dentist
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Dentist to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/domain.Dentist'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
      summary: Post Dentist
      tags:
      - Dentist
  /dentists/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Dentist
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Dentist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
      summary: Delete Dentist
      tags:
      - Dentist
    get:
      consumes:
      - application/json
      description: Get Dentist
      parameters:
      - description: Dentist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.response'
      summary: Get Dentist by ID
      tags:
      - Dentist
    patch:
      consumes:
      - application/json
      description: Patch a Dentist
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Dentist to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/domain.Dentist'
      - description: Dentist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.response'
      summary: Patch Dentist
      tags:
      - Dentist
    put:
      consumes:
      - application/json
      description: Put a Dentist
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Dentist to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/domain.Dentist'
      - description: Dentist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.response'
      summary: Put Dentist
      tags:
      - Dentist
  /patients:
    get:
      consumes:
      - application/json
      description: Get Patient
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
      summary: Get All Patient
      tags:
      - Patient
    post:
      consumes:
      - application/json
      description: Post a Patient
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Patient to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/domain.Patient'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
      summary: Post Patient
      tags:
      - Patient
  /patients/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Patient
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
      summary: Delete Patient
      tags:
      - Patient
    get:
      consumes:
      - application/json
      description: Get Patient
      parameters:
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.response'
      summary: Get Patient by ID
      tags:
      - Patient
    patch:
      consumes:
      - application/json
      description: Patch a Patient
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Patient to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/domain.Patient'
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.response'
      summary: Patch Patient
      tags:
      - Patient
    put:
      consumes:
      - application/json
      description: Put a Patient
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Patient to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/domain.Patient'
      - description: Patient ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.response'
      summary: Put Patient
      tags:
      - Patient
  /appointments:
    get:
      consumes:
      - application/json
      description: Get appointment
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
      summary: Get All appointment
      tags:
      - appointment
    post:
      consumes:
      - application/json
      description: Post a appointment
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: appointment to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/dto.appointmentInsert'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
      summary: Post appointment
      tags:
      - appointment
  /appointments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a appointment
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
      summary: Delete appointment
      tags:
      - appointment
    get:
      consumes:
      - application/json
      description: Get appointment
      parameters:
      - description: appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.response'
      summary: Get appointment by ID
      tags:
      - appointment
    patch:
      consumes:
      - application/json
      description: Patch a appointment
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: appointment to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/dto.appointmentInsert'
      - description: appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.response'
      summary: Patch appointment
      tags:
      - appointment
    put:
      consumes:
      - application/json
      description: Put a appointment
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: appointment to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/dto.appointmentInsert'
      - description: appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.response'
      summary: Put appointment
      tags:
      - appointment
  /appointments/dni:
    get:
      consumes:
      - application/json
      description: Get appointment
      parameters:
      - description: DNI
        in: query
        name: dni
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.response'
      summary: Get appointment by ID
      tags:
      - appointment
  /appointments/post:
    post:
      consumes:
      - application/json
      description: Post a appointment
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: appointment to data
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/dto.appointmentPost'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.response'
      summary: PostWithLicenseAndDNI appointment
      tags:
      - appointment
swagger: "2.0"